{"schema":"decodex/commit/1","summary":"Move replay summary finalization","authority":"manual"}
